add sql viewer for query
