+++ packages/legend-graph/src/graph/helpers/STO_Relational_Helper.ts
@@ -32,6 +32,26 @@ import type { Schema } from '../metamodel/pure/packageableElements/store/relatio
 import type { Table } from '../metamodel/pure/packageableElements/store/relational/model/Table.js';
 import type { View } from '../metamodel/pure/packageableElements/store/relational/model/View.js';
 import { RelationalInputType } from '../metamodel/pure/packageableElements/mapping/DEPRECATED__MappingTest.js';
+import {
+  type RelationalDataType,
+  VarChar,
+  Char,
+  VarBinary,
+  Binary,
+  Bit,
+  Numeric,
+  Decimal,
+  Double,
+  Float,
+  Real,
+  Integer,
+  SmallInt,
+  TinyInt,
+  Timestamp,
+  Other,
+  SemiStructured,
+  Json,
+} from '../../STO_Relational_Exports.js';
 
 const collectIncludedDatabases = (
   results: Set<Database>,
@@ -161,3 +181,46 @@ export const getColumn = (relation: Table | View, name: string): Column =>
       .find((column) => column.name === name),
     `Can't find column '${name}' in table '${relation.name}'`,
   );
+
+export const stringifyDataType = (type: RelationalDataType): string => {
+  if (type instanceof VarChar) {
+    return `VARCHAR(${type.size})`;
+  } else if (type instanceof Char) {
+    return `CHAR(${type.size})`;
+  } else if (type instanceof VarBinary) {
+    return `VARBINARY(${type.size})`;
+  } else if (type instanceof Binary) {
+    return `BINARY(${type.size})`;
+  } else if (type instanceof Bit) {
+    return `BIT`;
+  } else if (type instanceof Numeric) {
+    return `NUMERIC(${type.precision},${type.scale})`;
+  } else if (type instanceof Decimal) {
+    return `DECIMAL(${type.precision},${type.scale})`;
+  } else if (type instanceof Double) {
+    return `DOUBLE`;
+  } else if (type instanceof Float) {
+    return `FLOAT`;
+  } else if (type instanceof Real) {
+    return `REAL`;
+  } else if (type instanceof Integer) {
+    return `INTEGER`;
+  } else if (type instanceof BigInt) {
+    return `BIGINT`;
+  } else if (type instanceof SmallInt) {
+    return `SMALLINT`;
+  } else if (type instanceof TinyInt) {
+    return `TINYINT`;
+  } else if (type instanceof Date) {
+    return `DATE`;
+  } else if (type instanceof Timestamp) {
+    return `TIMESTAMP`;
+  } else if (type instanceof Other) {
+    return `OTHER`;
+  } else if (type instanceof SemiStructured) {
+    return 'SEMI-STRUCTURED';
+  } else if (type instanceof Json) {
+    return 'JSON';
+  }
+  return '(UNKNOWN)';
+};
